4-1 fix

+++ DB/overcloud/Models/CloudStorageInfo.cs
@@ -9,7 +9,7 @@ namespace DB.overcloud.Models
         public string AccountId { get; set; }            // 클라우드 로그인 ID
         public string AccountPassword { get; set; }      // 클라우드 로그인 비밀번호 (필요시)
 
-        public ulong TotalSize { get; set; }             // 계정의 전체 용량
-        public ulong UsedSize { get; set; }              // 계정의 사용 중인 용량
+        public int TotalSize { get; set; }             // 계정의 전체 용량
+        public int UsedSize { get; set; }              // 계정의 사용 중인 용량
     }
 }
